requirements.txt
streamlit==0.87.0
langchain==0.5.0
openai==0.10.2
PyPDF2==1.26.0
streamlit-extras==0.2.0
python-dotenv==0.19.2
streamlit-loading-animation==0.2.0
